Automatically try a restart if we think we've converged
This should cover cases where dphi < 0 "by a hair," i.e., where the search direction is almost orthogonal to the gradient. It's an extra layer of security beyond commit 50b9d60a9788798e68f010cd5a84d2b89626e582 to increase the likelihood that when we terminate, we really have reached a minimum.