Support multi-parameter α in linesearch, check slope during linesearch, and fix quadratic parameters
Since the step size for the inequality λ may not be the same as the step size for the rest of the parameters, we need a more sophisticated model of the function's behavior during linesearch. This commit implements a 4-parameter vector (αx, αs, αI, αE) for changes to the position, slack, inequality LM, and equality LM respectively (LM=lagrange multiplier). Consequently the "slope" becomes a 4-parameter vector and the "curvature" a 4x4 matrix.


In practice this helps convergence on some "hard" problems (e.g., HATFLDF).