Enable AppVeyor
Optim.jl has failures on Windows which are undetected by the devs because Windows CI is not enabled. As a large package in the Julia package ecosystem which has no binary dependencies, it should be easy to test and support Windows. This PR adds an appveyor.yml file to do just that. I would be willing to take the CI burden if you want (Orgs cannot have CI for AppVeyor).