Fix a use-before-initialisation bug
After finding the solution, Optim.optimize() assembles some information, which
includes comparing the solution to the previous iteration.  However, if the
solution was found in one iteration (i.e. the initial guess was right), then
there is no previous guess to compare with.  In this case, Optim.optimize()
crashes, because it tries to use something that has not been initialised.

Here is a simple example that triggers a crash:

	using Optim
	f(x) = x[1]^2
	x0 = BigFloat[0]
	obj = OnceDifferentiable(f, x0; autodiff=:forward)
	result = Optim.optimize(f, x0, BFGS())