Conv. criteria: add per-component g_abstol
This allows `options.g_abstol` to be either a scalar (the traditional
choice) or a vector. The motivation for this change comes from the fact
that the gradient is dependent on how each variable is scaled:

     f₁(x, y) = x^2 + y^2 + 1

and

     f₂(x, y) = x^2 + 10^8 * y^2 + 1

Both have the same minimizer and minimum value, but at any point other
than the x-axis, derivatives with respect to `y` of `f₂` are much larger
than those of `f₁`. When roundoff error becomes a factor, the
convergence requirement should be adjusted accordingly.